+++ resources/plugin.rb
@@ -26,9 +26,11 @@ action :install do
   plugin_name = new_resource.name
   binary = new_resource.grafana_cli_bin
   raise "#{plugin_name} is not available" unless ::GrafanaCookbook::Plugin.available?(plugin_name, binary)
+  service 'grafana-server'
   execute "Installing plugin #{plugin_name}" do
     command ::GrafanaCookbook::Plugin.build_cli_cmd(plugin_name, 'install', binary)
     not_if { GrafanaCookbook::Plugin.installed?(new_resource.name, new_resource.grafana_cli_bin) }
+    notifies :restart, 'service[grafana-server]'
   end
 end
 
@@ -36,8 +38,10 @@ action :update do
   plugin_name = new_resource.name
   binary = new_resource.grafana_cli_bin
   if GrafanaCookbook::Plugin.installed?(new_resource.name, new_resource.grafana_cli_bin)
+    service 'grafana-server'
     execute "Updating plugin #{plugin_name}" do
       command ::GrafanaCookbook::Plugin.build_cli_cmd(plugin_name, 'update', binary)
+      notifies :restart, 'service[grafana-server]'
     end
   else
     Chef::Log.warn "Impossible to upgrade plugin #{plugin_name} because it is not installed. We will install it."
@@ -48,8 +52,10 @@ end
 action :remove do
   plugin_name = new_resource.name
   binary = new_resource.grafana_cli_bin
+  service 'grafana-server'
   execute "Removing plugin #{name}" do
     command ::GrafanaCookbook::Plugin.build_cli_cmd(plugin_name, 'remove', binary)
     only_if { GrafanaCookbook::Plugin.installed?(new_resource.name, new_resource.grafana_cli_bin) }
+    notifies :restart, 'service[grafana-server]'
   end
 end